4.21 tehty

+++ osa4/blogilista/controllers/blogRouter.js
@@ -10,15 +10,13 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
-  const token = request.token
-
   try{
-    const decodedToken = await jwt.verify(token, config.SECRET)
-    if(!token || !decodedToken.id) {
+    const decodedToken = await jwt.verify(request.token, config.SECRET)
+    if(!request.token || !decodedToken.id) {
       response.status(400).send({error: 'bad username or password'})
       return
     }
-    
+
     const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
@@ -42,8 +40,22 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = await jwt.verify(request.token, config.SECRET)
+    if(!request.token || !decodedToken.id) {
+      response.status(400).send({error: 'bad username or password'})
+      return
+    }
+
+    const blogToRemove = await Blog.findById(request.params.id)
+    console.log(blogToRemove)
+    console.log()
+    
+    if(blogToRemove.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(400).end()
+    }
   } catch (exception) {
     console.log(exception)
     response.status(404).end()
